fix: typescriptnya (semoga) dah bener
