+++ packages/api/src/router/admin.ts
@@ -1,7 +1,8 @@
 import type { TRPCRouterRecord } from "@trpc/server";
 import { TRPCError } from "@trpc/server";
 
-import { and, eq, not, schema, sql } from "@sora-vp/db";
+import { and, eq, not, sql } from "@sora-vp/db";
+import schema from "@sora-vp/db/schema";
 import { admin } from "@sora-vp/validators";
 
 import { adminProcedure } from "../trpc";
