+++ packages/api/src/router/auth.ts
@@ -2,7 +2,8 @@ import type { TRPCRouterRecord } from "@trpc/server";
 import { TRPCError } from "@trpc/server";
 import bcrypt from "bcrypt";
 
-import { countUserTable, preparedGetUserByEmail, schema } from "@sora-vp/db";
+import { countUserTable, preparedGetUserByEmail } from "@sora-vp/db/client";
+import schema from "@sora-vp/db/schema";
 import { auth as authValidator } from "@sora-vp/validators";
 
 import { publicProcedure } from "../trpc";
