+++ tooling/typescript/internal-package.json
@@ -5,7 +5,7 @@
     /** Emit types for internal packages to speed up editor performance. */
     "declaration": true,
     "declarationMap": true,
-    "noEmit": false,
-    "emitDeclarationOnly": true
+    "emitDeclarationOnly": true,
+    "noEmit": false
   }
 }
